+++ LsHserver/src/main/scala/eu/brosbit/immovable/Plankton.scala
@@ -8,8 +8,8 @@ object Plankton extends PlantConst {
   def apply() = new Plankton
   override val name: String = "Plankton"
   override val shortName: String = "pk"
-  override val maxBio: Short = 100
-  override val growBio: Short = 15
-  override val growHP: Short = 0
-  override val maxHP: Short = 0
+  override val maxBio: Int = 100
+  override val growBio: Int = 15
+  override val growHP: Int = 0
+  override val maxHP: Int = 0
 }
